+++ sdk-base/src/main/java/com/mapbox/maps/MapLoadingErrorDelegate.kt
@@ -1,0 +1,17 @@
+package com.mapbox.maps
+
+import androidx.annotation.RestrictTo
+
+/**
+ * For internal usage.
+ *
+ * @suppress
+ */
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun interface MapLoadingErrorDelegate {
+
+  /**
+   * Send [MapLoadingError].
+   */
+  fun sendMapLoadingError(error: MapLoadingError)
+}